==> 1. Crime data cleaning.ipynb <==
import pandas as pd
import fnmatch
import numpy as np
from mlxtend.preprocessing import TransactionEncoder
from mlxtend.frequent_patterns import apriori
from mlxtend.frequent_patterns import association_rules
import matplotlib.pyplot as plt 

# Load the crime data from Brottsförebyggande Rådet (BRÅ).
df = pd.read_excel("crime_data_bra.xlsx", engine='openpyxl')

# Drop unecessary columns.
df = df.drop(["Brott"], axis=1)
df

# Drop first row (only NaN value).
df = df.iloc[1:]

# Translate date string into int months.

# Use dictionary to translate months.
m = {
    "Jan": "1",
    "Feb": "2",
    "Mar": "3",
    "Apr": "4",
    "Maj": "5",
    "Jun":"6",
    "Jul":"7",
    "Aug":"8",
    "Sep":"9",
    "Okt":"10",
    "Nov":"11",
    "Dec": "12"
}

new_months = []
for month in df["Period"]:
    new_months.append(m[month])
df["Period"] = new_months
df["År"] = df["År"].apply(lambda x: int(x))

# Save the date in a combined column called 'date'.
# The format of this column should be YYYY-MM.

new_col = []
for a,p in zip(df["År"], df["Period"]):
    if int(p) < 10:
        new_col.append(str(a)+'-'+'0'+str(p))
    else:
        new_col.append(str(a)+'-'+str(p))
df["Date"]= new_col
df = df.drop(["År", "Period"], axis=1)
df.sort_values(by="Date")

# Check that the date is now formatted correctly.
df

# Save the new cleaned dataset to a CSV file.
df.to_csv("crime_data_cleaned.csv")
