==> 3. Join weather data with regions.ipynb <==
import pandas as pd
from IPython.display import clear_output

# Retrieve the region data
region_data = pd.read_csv("name_region_data.csv")
region_data = region_data[["Name", "County"]]
region_data["County"][3] = "Västra Götalands län"

# Retrieve the weather data
weather_data = pd.read_csv("weather_data.csv")

# Get the locations and set as same
locations = weather_data["location"]
set(region_data["Name"]) - set(locations)

# Get weather data from 1995 and later
weather_data_from_1995 = weather_data[weather_data["date"]>="1995-01-01"]
weather_data_from_1995

# Add the county from the region data to each weather observation.
def add_regions(df, name_regions):
    new_col = []
    count = 0
    for region in df["location"]:
        count = count +1
        try:
            new_value = list(name_regions[name_regions["Name"]==region]["County"])[0]
            new_col.append(new_value)
        except KeyError:
            new_col.append("Error")
        finally:
            clear_output(wait=True)
            print(str(count) + " / "+str(len(df["location"]))  )
    df["Region"] = new_col
    return df

weather_regions = add_regions(weather_data_from_1995, region_data)

# Save the weather data with county to a CSV file.
weather_regions.to_csv("weather_regions_new.csv")

weather_regions
