==> 4. Merge weather and crime data.ipynb <==
import pandas as pd
import fnmatch
import numpy as np

# Import the cleaned crime data
crime_df = pd.read_csv("crime_data_cleaned.csv").drop("Unnamed: 0", axis=1)
crime_df

# Import the weather data with county information.
weather_regions_df = pd.read_csv("weather_county_data.csv").drop("Unnamed: 0", axis=1)
weather_regions_df

# Update names of columns on weather data & remove days from date
weather_regions_df.columns = ["Location", "Date", "Snow_depth", "Region"]
weather_regions_df["Date"] = weather_regions_df["Date"].apply(lambda x: x[:-3])
weather_regions_df

# The weather data is coded with a number depending on how much snow is on the
# ground. We are only interested in days when there is snow so we can
# disregard the other days. Days with snow is coded 1, no snow is coded 0.

def snow_day(x):
    if (x == 17) | (x == 14)|(x == 17) | (x == 18) | (x == 19):
        return 1
    else:
        return 0

# We can now drop unneccecary columns.
weather_regions_df = weather_regions_df.drop("Snow_depth", axis=1)

# Group and sum the number of snow days.
weather_regions_snow_days = weather_regions_df.groupby(["Location", "Date", "Region"]).sum()

weather_regions_snow_days.reset_index(inplace=True)
weather_regions_snow_days

# Create the final merged dataset by merging the crime and
# snow dataframes.
merged_df = crime_df.merge(weather_regions_snow_days, on=["Region", "Date"])
merged_df

# Now when the data has been merged we now need to aggregate the number of
# snow days and crimes each month. The following script goes through the 
# data and creates the artifical number 'snow_count'. Snow count
# is the number of snow observations divided by the number of stations.
# The script also aggregates the number of crimes.

prev_date = ""
prev_region = ""
snow_count = 0
crime_count = ""
stations_count = 0
max_index = merged_df.shape[0]-1

data = []

for i, row in merged_df.iterrows():
    
    curr_region = row['Region']
    curr_date = row['Date']
    crime = row['Antal']
    snow = row['Snow_days']
    
    if(i == 0):
        prev_date = curr_date
        prev_region = curr_region
        stations_count = 1
    
    if (prev_date != curr_date or prev_region != curr_region):
        
        converted_date = pd.to_datetime(prev_date, format='%Y-%m', errors='ignore')
        to_append = [prev_region, converted_date, crime_count, (snow_count / stations_count)]
        data.append(to_append)
        
        prev_date = curr_date
        prev_region = curr_region
        
        snow_count = 0
        stations_count = 1
        crime_count = crime
        snow_count = snow
    elif (i == max_index):
        to_append = [prev_region, prev_date, crime_count, (snow_count / stations_count)]
        data.append(to_append)
    else:
        crime_count = crime
        snow_count = snow_count + snow
        stations_count = stations_count + 1      
        
aggregated_df = pd.DataFrame(data, columns=['region', 'date', 'crime_count', 'snow_count'])

# Create a new index for crime count.
index_names = aggregated_df[ aggregated_df['crime_count'] == ".." ].index
  
# Drop row indexes from dataFrame
aggregated_df.drop(index_names, inplace = True)

# Convert numbers to integers
aggregated_df["crime_count"] = aggregated_df["crime_count"].apply(lambda x: float(x))
aggregated_df

aggregated_df.to_csv("final_dataset.csv")
